requirements.txt
langchain
langgraph
langgraph-prebuilt
langchain-groq
langchain-ollama
langchain-openai
langchain-anthropic
langchain-community
langchain-chroma
langchain-huggingface
langchain-text-splitters
langchain-mcp-adapters
chromadb
sentence-transformers
fastmcp
mcp
rich
python-dotenv
tavily-python
